Add rounding for text drawing and translation to fix font glitch
Drawing fonts at non-integer positions sometimes leads to glyphs not
being rendered correctly.
This commit attempts to fix this by flooring the position at which the
text is drawn, as well as all translations. This is somewhat of a hack
since flooring all translations might lead to an accumulative rounding
error over time. For example translating by 2.5 px and then by .5 px is
not the same as translating by 3 px