Export quads in predictable order
Previously the exporter would just iterate over the table's value and
write them out as they come in. That would lead to different results
each time the table was exported, even if the table's content didn't
change.

This commit sorts the table's keys, and writes out the table's content
in alphabetical order.