Add file to recent files whenever it is loaded, not just when it is picked
Previously a file would only be added to recent files when it was picked from
the file browser. Hoever, we also want to update the list of recent files when
the file is loaded from somewhere else. With this commit the list of recent
files is updated whenever a quad file is loaded, no matter how