Remove auto-detect feature for .qua file
Since the image is stored in tue qua file, it is faster to drag the image
into the window or load it via the file menu or recent files.
Also, the hacky function that was used to find qua files next to the image
is not able to cover cases where image and qua file are in different
directories.