Shorten APPVERSION if current commit is tagged commit
Using git describe, we automatically get e.g. 0.5.0 if there are no commits
between HEAD and the commit 0.5.0, and we get something like 0.5.0-2-2ff4b6ca
if HEAD is at 2ff4b6ca and there has been two commits since 0.5.0.

We use this instead of the previous, longer versions so that the released
versions (which should always be tagged) show nice, easy to parse version
numbers, while in-development version still show everything that is necessary
to know which version is used.