Replace update_license by check_license to minimize unnecessary work
When we update the license each time we make distfiles, then a large portion
of the distfiles will be recreated each time, since Quadtastic/libquadtastic.lua
is updated whenever update_license is made.

This commit adds a recipe that simply checks whether the licenses are up to date
and errors if they aren't. Since this will only happen about once a year, this
is preferable to recomputing a bunch of files each time.

The update_license recipe is still present and can be triggered manually.