Add Exporters, menu entry and WIP readme
The Exporters module exposes functions to list all custom exporters.
The menu entry allows the user to use a custom exporter, to open the directory
that contains the custom exporters, and to refresh the list of custom exporters.

The WIP readme will soon give a short introduction to how to write custom
exporters.