Change separation of functions in exporters module
The previous commit extracted all functionality for loading a module from
a file and checking it for the mandatory fields and functions that define
whether it is a valid exporter. However, what I actually need is a function
that checks whether a module is an exporter without any dependencies on love
functionality. So I redid the separation, which also lead to a cleaner
functionality scope.