Add README for exporters
This README is in the source tree and not in the wiki because GitHub does not
really support linking the source tree to a specific version of the wiki.
Anyone who would want to contribute an exporter would need to edit the wiki,
and those changes would be live immediately.

By keeping this piece of documentation in the source tree, it can be properly
altered on a separate branch for a pull request.