When displaying files, always display them in alphabetical order
So far, the order in which files were displayed in the file dialogs depended on
the order in which they were enumerated by the file system. This commit
explicitly sorts the filenames alphabetically and case-insensitive, so that the
order in which they are displayed is not dependend on the used file system.